UNG-784 Adding some class organization to better observe v1 and v2
